Move controllers to their own package in preparation for API controllers.
